Updated doPost method to reflect parameter change in filterByTicketNr

+++ src/main/parkhaus-project/Parkhaus/java/com/se1_team20/Parkhaus/CHECKOUT/CheckoutAuthenticationServlet.java
@@ -45,7 +45,7 @@ public abstract class CheckoutAuthenticationServlet extends ParkingServlet {
     {
         response.setContentType("text/html");
         final CheckoutModel MODEL = new CheckoutModel();
-        final Double enteredTicketNR  = Double.parseDouble(request.getParameter("entered_ticket_nr"));
+        final Integer enteredTicketNR  = Integer.parseInt(request.getParameter("entered_ticket_nr"));
         final List<CarIF> CARS           = (List<CarIF>) request.getServletContext().getAttribute("former-cars" + "Level1");
         final PrintWriter OUT               = response.getWriter();
         OUT.write("<html><body><div id='servletResponse' style='text-align: center;'>");
